requirements.txt
Flask
Flask-WTF
Jinja2
Werkzeug
Flask-Login
boto3
Flask-Mail
python-dotenv
